relayer: move BalanceMonitor into monitor
